-- 소슻취약점 수정

+++ src/main/java/egovframework/rte/ygpa/gam/cmmn/web/GamGroupMngController.java
@@ -216,7 +216,7 @@ public class GamGroupMngController {
 
     	GroupManage gm = new GroupManage();
 //    	List list = (ArrayList)deleteList.get("delList");
-       if(delList.length > 0 ){
+       if(delList.length > 0 && gm !=null ){
     	for(int i=0; i<delList.length; i++) {
     		gm.setGroupId((String)delList[i]);
     		egovGroupManageService.deleteGroup(gm);
